+++ mmc/__init__.py
@@ -1,4 +1,4 @@
-from typing import List, Iterable, Tuple
+from typing import List, Iterable, Tuple, Callable
 
 from api_session import APISession, JSONDict
 
@@ -6,11 +6,19 @@ __all__ = ["Mattermost", "__version__"]
 
 __version__ = "0.1.0"
 
+
 def ordered_items(response: JSONDict, items_keys: str) -> List[Tuple[str, JSONDict]]:
     return [
         (item_id, response[items_keys][item_id])
         for item_id in response["order"]
     ]
+
+
+def min_version(_version: str):
+    def decorator(fn: Callable):
+        return fn
+
+    return decorator
 
 
 class Mattermost(APISession):
@@ -49,6 +57,7 @@ class Mattermost(APISession):
         """Get all users."""
         return self._get_paginated_entities("/users", **kwargs)
 
+    @min_version("5.10")
     def get_bots(self, **kwargs) -> Iterable[JSONDict]:
         """Get all bots."""
         return self._get_paginated_entities("/bots", **kwargs)
@@ -65,6 +74,7 @@ class Mattermost(APISession):
             else:
                 return
 
+    @min_version("5.34")
     def get_team_files(self, team_id: str, terms: str) -> Iterable[JSONDict]:
         """Get files matching search terms."""
         resp = self.post_api(f"/teams/{team_id}/files/search", json={
@@ -82,3 +92,28 @@ class Mattermost(APISession):
     def delete_post(self, post_id: str):
         """Delete a post."""
         return self.delete_api(f"/posts/{post_id}", throw=True).json()
+
+    # Usage
+    # =====
+
+    def get_total_channels_count(self, include_deleted=False, **kwargs) -> int:
+        """Return the total channels count, for all teams."""
+        kwargs.setdefault("params", {})
+        kwargs["params"].update({
+            "per_page": 1,
+            "include_total_count": "true",
+            "include_deleted": "true" if include_deleted else "false",
+        })
+
+        payload = self.get_json_api("/channels", **kwargs)
+        return payload["total_count"]
+
+    @min_version("7.0")
+    def get_total_posts_count(self, **kwargs) -> int:
+        """Return the total posts count, for all teams."""
+        return self.get_json_api("/usage/posts", **kwargs)["count"]
+
+    @min_version("7.1")
+    def get_total_file_storage_usage_bytes(self, **kwargs) -> int:
+        """Return the total file storage usage, in bytes."""
+        return self.get_json_api("/usage/storage", **kwargs)["bytes"]